chore: replace Cpcdos branding to Munix
Hello everyone !
As you can see the Cpcdos kernel is now Munix. Why have I done that you
might ask ? The Cpcdos community doesn't have the knowledge to maintain
kernels. As of the main (and only) maintainer of the project, I give it
to our brand new project : Cute Engineering.